updated astro endpoint to use loaders/api/helpers
