+++ src/loaders/api/index.ts
@@ -6,10 +6,21 @@ import peopleLoader from "./people";
 import placesLoader from "./places";
 import worksLoader from "./works";
 
+const loaderDict = {
+  events: eventsLoader,
+  instances: instancesLoader,
+  items: itemsLoader,
+  organizations: organizationsLoader,
+  people: peopleLoader,
+  places: placesLoader,
+  works: worksLoader,
+};
+
 export {
   eventsLoader,
   instancesLoader,
   itemsLoader,
+  loaderDict,
   organizationsLoader,
   peopleLoader,
   placesLoader,
